+++ rectools/compat.py
@@ -68,3 +68,9 @@ class MetricsApp(RequirementUnavailable):
     """Dummy class, which is returned if there are no dependencies required for the model"""
 
     requirement = "visuals"
+
+
+class CatBoostReranker(RequirementUnavailable):
+    """Dummy class, which is returned if there are no dependencies required for the model"""
+
+    requirement = "catboost"
